Change file structure
