+++ packages/strapi-plugin-i18n/admin/src/middlewares/extendCTBAttributeInitialDataMiddleware.js
@@ -1,35 +1,3 @@
-import { has } from 'lodash';
-
-const extendCTBInitialDataMiddleware = () => {
-  return () => next => action => {
-    if (
-      action.type === 'ContentTypeBuilder/FormModal/SET_DATA_TO_EDIT' &&
-      action.modalType === 'contentType'
-    ) {
-      const i18n = { localized: false };
-
-      const pluginOptions = action.data.pluginOptions
-        ? { ...action.data.pluginOptions, i18n }
-        : { i18n };
-
-      const data = { ...action.data, pluginOptions };
-
-      if (action.actionType === 'create') {
-        return next({ ...action, data });
-      }
-
-      // Override the action if the pluginOption config does not contain i18n
-      // In this case we need to set the proper initialData shape
-      if (!has(action.data.pluginOptions, 'i18n.localized')) {
-        return next({ ...action, data });
-      }
-    }
-
-    // action is not the one we want to override
-    return next(action);
-  };
-};
-
 const extendCTBAttributeInitialDataMiddleware = () => {
   return ({ getState }) => next => action => {
     const enhanceAction = () => {
@@ -86,6 +54,4 @@ const extendCTBAttributeInitialDataMiddleware = () => {
   };
 };
 
-const middlewares = [extendCTBInitialDataMiddleware, extendCTBAttributeInitialDataMiddleware];
-
-export default middlewares;
+export default extendCTBAttributeInitialDataMiddleware;
