+++ packages/strapi-plugin-i18n/admin/src/middlewares/index.js
@@ -1,0 +1,6 @@
+import extendCTBInitialDataMiddleware from './extendCTBInitialDataMiddleware';
+import extendCTBAttributeInitialDataMiddleware from './extendCTBAttributeInitialDataMiddleware';
+
+const middlewares = [extendCTBInitialDataMiddleware, extendCTBAttributeInitialDataMiddleware];
+
+export default middlewares;
